requirements.txt
asgiref==3.5.0
Django==3.2.12
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-jazzmin==2.4.9
django-js-asset==2.0.0
django-multiselectfield==0.1.12
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
